+++ src/types/metadata.d.ts
@@ -1,0 +1,7 @@
+export interface Metadata {
+  title: string;
+  hvtID: string;
+  author: string;
+  affiliation: string;
+  publicationDate: string;
+}
